- Moving our script to match what we're doing in other Azure repos (specifically, azure-rest-api-specs). - Adding a comment into the script,. in case someone finds it first and isn't sure how it fits into the picture. - Removing the main attribute. We're a package, but not really, and that's okay.
